+++ src/index.ts
@@ -1,3 +1,5 @@
+#!/usr/bin/env node
+/* eslint-disable node/shebang */
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import * as pkg from "../package.json";